# storage/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is swagger documentation for store store",
    "version": "0.0.1",
    "title": "Store Backend",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "miqdad.farcha@store.backend.fdc"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "host": "store.backend.fdc",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/stats/daily-sales": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Get Stat daily Sales",
        "description": "",
        "operationId": "getDailySales",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "The Date",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              },
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              }
            },
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Invalid data supplied"
          },
          "401": {
            "description": "Unauthorized access"
          }
        }
      }
    },
...
  },
...
}